feat : change password

+++ src/services/AppContentRoute.jsx
@@ -1,5 +1,11 @@
 import { useEffect } from "react";
-import { Routes, Route, useLocation, useNavigate } from "react-router-dom";
+import {
+  Routes,
+  Route,
+  useLocation,
+  useNavigate,
+  useSearchParams,
+} from "react-router-dom";
 import { isAuthenticated } from "../utils/auth";
 import { LocalizationProvider } from "@mui/x-date-pickers";
 import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs";
@@ -7,10 +13,12 @@ import { LoaderProvider } from "../utils/context/LoaderContext";
 import LoginPage from "../components/LoginPage";
 import AdminLayouts from "../components/layouts/AdminLayouts";
 import EmployeeLayouts from "../components/layouts/EmployeeLayouts";
+import ForgotPasswordPage from "../components/ForgotPasswordPage";
 
 export const AppContentRoute = () => {
-  //const location = useLocation();
+  const location = useLocation();
   const navigate = useNavigate();
+  const [searchParams, setSearchParams] = useSearchParams();
   // const isLoginPage = location.pathname === "/login";
   const { role } = isAuthenticated();
 
@@ -31,6 +39,19 @@ export const AppContentRoute = () => {
     return false;
   };
   useEffect(() => {
+    console.log(location.pathname);
+    if (location.pathname === "/change-password") {
+      console.log(searchParams.get("email"));
+      console.log(searchParams.get("token"));
+      const resetEmail = searchParams.get("email");
+      const resetToken = searchParams.get("token");
+      if (resetEmail && resetToken) {
+        localStorage.setItem("resetEmail", resetEmail);
+        localStorage.setItem("resetToken", resetToken);
+      }
+      navigate("/change-password");
+      return;
+    }
     if (!isValidLoggedIn()) {
       navigate("/login");
     }
@@ -38,7 +59,7 @@ export const AppContentRoute = () => {
 
   const getRoleBasedLayout = () => {
     // To access role , user must have valid login
-    if (!isValidLoggedIn()) {
+    if (!isValidLoggedIn() && location.pathname != "/change-password") {
       localStorage.clear();
     }
     switch (role.toLowerCase()) {
@@ -47,7 +68,11 @@ export const AppContentRoute = () => {
       case "employee":
         return <EmployeeLayouts />;
       default:
-        return <LoginPage />;
+        return location.pathname == "/change-password" ? (
+          <ForgotPasswordPage />
+        ) : (
+          <LoginPage />
+        );
     }
   };
